+++ src/main/kotlin/de/bund/digitalservice/useid/apikeys/ApiKeyDetails.kt
@@ -1,0 +1,3 @@
+package de.bund.digitalservice.useid.apikeys
+
+class ApiKeyDetails(val keyValue: String, val refreshAddress: String?, val requestDataGroups: List<String>)
