+++ src/main/kotlin/de/bund/digitalservice/useid/config/CustomWebClientExchangeTagsProvider.kt
@@ -1,10 +1,10 @@
 package de.bund.digitalservice.useid.config
 
 import io.micrometer.core.instrument.Tag
+import jakarta.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletResponse
 import org.springframework.boot.actuate.metrics.web.servlet.WebMvcTagsContributor
 import org.springframework.stereotype.Component
-import javax.servlet.http.HttpServletRequest
-import javax.servlet.http.HttpServletResponse
 
 @Component
 class CustomWebClientExchangeTagsProvider : WebMvcTagsContributor {
