+++ src/main/kotlin/de/bund/digitalservice/useid/config/TenantProperties.kt
@@ -22,18 +22,4 @@ class TenantProperties {
     fun findByTenantId(id: String?): Tenant? {
         return tenants.find { it.id == id }
     }
-
-    class Tenant {
-        lateinit var id: String
-
-        // api key information
-        lateinit var apiKey: String
-        lateinit var refreshAddress: String
-        var dataGroups: List<String> = emptyList()
-
-        // csp header information
-        lateinit var defaultConfig: String
-        lateinit var frameAncestors: String
-        lateinit var allowedHost: String
-    }
 }
